pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perceptra-zero-shot"
version = "0.1.0"
description = "Production-ready zero-shot object detection and segmentation toolkit"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Perceptra Team", email = "team@perceptra.ai"}
]
keywords = [
    "computer-vision",
    "object-detection",
    "zero-shot",
    "deep-learning",
    "vision-language-models"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0,<2.0.0",
    "opencv-python>=4.5.0",
    "scipy>=1.7.0",
    "timm>=0.9.0",
    "huggingface-hub>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
]

api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

cli = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
]

viz = [
    "matplotlib>=3.5.0",
]

all = [
    "perceptra-zero-shot[dev,api,cli,viz]",
]

[project.scripts]
perceptra-zero-shot = "perceptra_zero_shot.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/perceptra-zero-shot"
Documentation = "https://perceptra-zero-shot.readthedocs.io"
Repository = "https://github.com/yourusername/perceptra-zero-shot"
Issues = "https://github.com/yourusername/perceptra-zero-shot/issues"

[tool.setuptools.packages.find]
include = ["perceptra_zero_shot*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=perceptra_zero_shot --cov-report=html --cov-report=term"